requirements.txt
langchain>=0.3.25
langchain-openai>=0.3.16
langchain-community>=0.3.23
chromadb>=1.0.8
sentence-transformers>=4.1.0
pypdf>=5.5.0
tiktoken>=0.9.0
numpy>=2.2.5
torch>=2.7.0
transformers>=4.51.3
openai>=1.78.0
python-dotenv>=1.1.0